# packages/server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/items":{"get":{"operationId":"getItems","summary":"Get items","parameters":[],"responses":{"200":{"description":""}},"tags":["Items"],"security":[{"bearer":[]}]}},"/api/auth/sign-up":{"post":{"operationId":"signUp","summary":"Sign up","description":"Sign up a new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/sign-in":{"post":{"operationId":"signIn","summary":"Sign in","description":"Sign in an existing user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}}},"tags":["Auth"]}},"/api/users/me":{"get":{"operationId":"getMyProfile","summary":"Get your user profile","description":"Gets the current user profile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/verification":{"post":{"operationId":"verifyUser","summary":"Verify User","description":"Verify a User using the token emailed to them during account creation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerificationDto"}}}},"responses":{"200":{"description":""},"401":{"description":""}},"tags":["Verification"]}}},"info":{"title":"Usufruct","description":"API Documentation","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"displayName":{"type":"string"}},"required":["email","password","displayName"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"displayName":{"type":"string"}},"required":["id","email","displayName"]},"VerificationDto":{"type":"object","properties":{"email":{"type":"string"},"token":{"type":"string"}},"required":["email","token"]}}}}